Add test runner and automate problem setup

+++ sliding-window/find_length.cpp
@@ -1,4 +1,5 @@
 #include <vector>
+#include <algorithm>
 
 using namespace std;
 
@@ -6,7 +7,7 @@ int findLength(vector<int> nums,int k) {
 
     int left = 0, curr = 0, ans = 0;
 
-    for (int right = 0; right < nums.size(); ++right) {
+    for (size_t right = 0; right < nums.size(); ++right) {
 
         curr += nums[right];
 
@@ -15,7 +16,7 @@ int findLength(vector<int> nums,int k) {
             left++;
         }
 
-        ans = max(ans, right - left + 1);
+        ans = max(ans, static_cast<int>(right - left + 1));
 
     }
